# api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Visky API
  version: 1.0.0
  description: VK Music Express API
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://visky.envarg.com
    description: Production server
...
paths:
  /auth/vk:
    get:
      description: Authorize user using VK auth page.
      summary: >
        Get VK auth page and remove defended scripts
        to fetch after auth token and secret. 
        Try to emulate old android app.
      operationId: getVkAuthPage
      tags:
        - authForm
      responses:
        '200':
          description: OK
          content:
            text/html:
              schema:
                type: string
                example: '<html><body>VK Auth Page</body></html>'
    post:
      summary: >
        get token and secret by login and password
        emulate old android app
      operationId: authorizeAndGetAuthToken
      tags:
        - authForm
      responses:
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errMessage:
                    type: string
                    example: 'Response data not contain secret or access_token'
        '301':
          description: Redirect to ehe /auth/blank.html
...
  /api/auth/token:
    post:
      description: Create user token.
      summary: Create new auth token with data from /auth/blank.html
      operationId: createAuthToken
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '200':
          description: Return auth token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthFragmentResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errMessage:
                    type: string
                example:
                  errMessage: "No 'access_token' or 'secret' in vk url"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errMessage:
                    type: string
                example:
                  errMessage: "Some API error occurred"
  /api/auth/refresh:
    post:
      description: Set data to session.
      summary: Set data to session and refresh existing token
      operationId: setAuthDataToSession
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  errMessage:
                    type: string
                example:
                  errMessage: "No `session` in request"
        '301':
          description: Redirect to ehe /auth/refresh
    get:
      security:
        - SessionCookieAuth: []
        - HeaderTokenAuth: []
      description: Refresh existing token.
      summary: get data from session and refresh auth token
      operationId: refreshAuthToken
      tags:
        - auth
      responses:
        '200':
          description: Return auth token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthFragmentResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  errMessage:
                    type: string
                    example: "No 'access_token' or 'secret' in vk url"
  /api/auth/profile:
    get:
      security:
        - SessionCookieAuth: []
        - HeaderTokenAuth: []
      description: Get user profile.
      summary: Get user profile information like name, id, etc.
      operationId: getUserProfile
      tags:
        - auth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/VkResponse'
                  - type: object
                    properties:
                      response:
                        $ref: '#/components/schemas/VkUserInfoResponse'
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: |
        Session authentification. 
        Client must send cookie with name 'connect.sid', received after login success.
    HeaderTokenAuth:
      type: apiKey
      in: header
      name: x-auth-token
      description: |
        Alternate authentification method. 
        Client must send a valid token in header with name 'x-auth-token'.
  schemas:
...
    TokenRequest:
      description: Token and secret
      type: object
      required:
        - secret
      properties:
        access_token:
          type: string
        url:
          type: string
        token:
          type: string
        secret:
          type: string
        user_id:
          type: string
        created:
          type: string
          format: date-time
          nullable: true
        maxAge:
          type: number
          nullable: true
        expires:
          type: string
          format: date-time
          nullable: true
...
    VkResponse:
      description: Response from VK
      type: object
      properties:
        response:
          oneOf:
            - $ref: '#/components/schemas/VkRefreshTokenResponse'
            - $ref: '#/components/schemas/VkPlaylistResponse'
            - $ref: '#/components/schemas/VkUserInfoResponse'
...
    VkUserInfoResponse:
      description: User info
      type: object
      properties:
        profile:
          type: object
          properties:
            id:
              type: integer
            first_name:
              type: string
            last_name:
              type: string
            can_access_closed:
              type: boolean
            is_closed:
              type: boolean
        info:
          type: object
          properties:
            2fa_required:
              type: integer
            audio_ads:
              type: object
              properties:
                day_limit:
                  type: integer
                track_limit:
                  type: integer
                types_allowed:
                  type: array
                  items:
                    type: string
                sections:
                  type: array
                  items:
                    type: string
            audio_permissions:
              type: object
              properties:
                audio_meta_info:
                  type: boolean
              example:
                audio_meta_info: false
            business_notify_enabled:
              type: integer
            change_email_url:
              type: string
            change_phone_url:
              type: string
            country:
              type: string
            debug_available:
              type: boolean
            email:
              type: string
            email_status:
              type: string
            eu_user:
              type: boolean
            feed_preloading:
              type: object
              properties:
                media_discover:
                  type: object
                  properties:
                    on_app_start:
                      type: object
                      properties:
                        enabled:
                          type: boolean
                breaking_news:
                  type: object
                  properties:
                    on_app_start:
                      type: object
                      properties:
                        enabled:
                          type: boolean
                    on_neighbour_tab:
                      type: object
                      properties:
                        enabled:
                          type: boolean
            https_required:
              type: integer
            intro:
              type: integer
            music_intro:
              type: integer
            fave_intro:
              type: integer
            menu_intro:
              type: boolean
            community_comments:
              type: boolean
            track_installed_apps:
              type: boolean
            clickable_stickers:
              type: object
              properties:
                max_stickers:
                  type: object
                  properties:
                    hashtag:
                      type: integer
                    mention:
                      type: integer
                    music:
                      type: integer
                    playlist:
                      type: integer
                    video:
                      type: integer
                    question:
                      type: integer
                    place:
                      type: integer
                    story_reply:
                      type: integer
                    owner:
                      type: integer
                    market_item:
                      type: integer
                    link:
                      type: integer
                    post:
                      type: integer
                    poll:
                      type: integer
                    sticker:
                      type: integer
                    app:
                      type: integer
                    situational_theme:
                      type: integer
                    spoiler:
                      type: integer
            reports_spa:
              type: boolean
            lang:
              type: integer
            money_p2p_params:
              type: object
              properties:
                min_amount:
                  type: integer
                max_amount:
                  type: integer
                currency:
                  type: string
                show_intro:
                  type: boolean
            no_wall_replies:
              type: integer
            own_posts_default:
              type: integer
            phone:
              type: string
            phone_status:
              type: string
            profiler_enabled:
              type: boolean
            profiler_settings:
              type: object
              properties:
                api_requests:
                  type: boolean
                download_patterns:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                      pattern:
                        type: string
                      probability:
                        type: number
                      error_probability:
                        type: integer
            raise_to_record_enabled:
              type: boolean
            settings:
              type: array
              items:
                type: object
                properties:
                  available:
                    type: boolean
                  forced:
                    type: boolean
                  name:
                    type: string
                  value:
                    type: string
            support_url:
              type: string
            valid_from:
              type: object
              properties:
                discover_posts:
                  type: integer
                discover_categories:
                  type: integer
            vk_pay_endpoint:
              type: string
            vk_pay_endpoint_v2:
              type: string
            stream_special_comment_price:
              type: integer
            menu_ads_easy_promote:
              type: object
              properties:
                item_url:
                  type: string
                item_text:
                  type: string
                show_badge:
                  type: boolean
            side_menu_custom_items:
              type: array
              items:
                type: object
                properties:
                  icon:
                    type: string
                  title:
                    type: string
                  action:
                    type: object
                    properties:
                      type:
                        type: string
                      target:
                        type: string
                      url:
                        type: string
            subscription_combo_allowed:
              type: boolean
            include_channel_notifications:
              type: boolean
            messages_transcript_auto_show:
              type: boolean
            messages_multiline_input:
              type: boolean
            messages_translation_language_pairs:
              type: array
              items:
                type: string
            obscene_text_filter:
              type: boolean
            messages_reaction_notifications:
              type: object
              properties:
                max_message_age_sec:
                  type: integer
                max_reactions:
                  type: integer
              required:
                - max_message_age_sec
                - max_reactions
            market_adult_18plus:
              type: object
              properties:
                is_adult_by_profile:
                  type: boolean
                is_adult_confirm:
                  type: boolean
              required:
                - is_adult_by_profile
                - is_adult_confirm
            can_change_password:
              type: boolean
            is_personal_ads_easy_promote_enabled:
              type: boolean
            messages_counter_settings:
              type: object
              properties:
                include_muted:
                  type: boolean
                include_group_dialogs:
                  type: boolean
                include_channels:
                  type: boolean
        time:
          type: integer
        has_new_items:
          type: object
          properties:
            global_promotion:
              type: integer
            store_new_items:
              type: integer
            stickers_version_hash:
              type: string
            favorite_stickers_version_hash:
              type: string
        allow_buy_votes:
          type: boolean
        ads_stoplist:
          type: array
        use_vigo:
          type: boolean
        html_games_supported:
          type: integer
        defaultAudioPlayer:
          type: string
      example:
        response:
          profile:
            id: 37758500
            first_name: "first_name"
            last_name: "last_name"
            can_access_closed: true
            is_closed: true
          info:
            2fa_required: 1,
            audio_ads:
              day_limit: 100,
              track_limit: 1,
              types_allowed:
                - "preroll"
              sections:
                - "my"
                - "user_playlists"
                - "group_playlists"
                - "my_playlists"
                - "recent"
                - "audio_feed"
                - "collections"
                - "genre_rock"
                - "genre_pop"
                - "genre_rap"
                - "artist_info"
                - "artist_main_albums"
                - "artist_main_audios"
                - "artist_featured_albums"
                - "artist_featured_audios"
                - "artist_featured_albums"
                - "artist_featured_albums"
                - "artist_pages"
                - "artist_events"
                - "artist_related"
                - "artist_new_album"
                - "artist_top_audios"
                - "artist_top_videos"
                - "artist_concerts"
            audio_permissions:
              audio_meta_info: false
            business_notify_enabled: 1
            change_email_url: "https://m.vk.com/settings?act=api_change_email&api_token=vk1.a.for_example--token"
            change_phone_url: "https://m.vk.com/activation?act=api_change_phone&api_token=vk1.a.for_example--token"
            country: "FI"
            debug_available: false
            email: "en***@gmail.com"
            email_status: "confirmed"
            eu_user: true
            feed_preloading:
              media_discover:
                on_app_start:
                  enabled: true
              breaking_news:
                on_app_start:
                  enabled: false
                on_neighbour_tab:
                  enabled: true
            https_required: 1
            intro: 0
            music_intro: 2
            fave_intro: 0
            menu_intro: true
            community_comments: false
            track_installed_apps: true
            clickable_stickers:
              max_stickers:
                hashtag: 4
                mention: 10
                music: 1
                playlist: 1
                video: 1
                question: 1
                place: 1
                story_reply: 1
                owner: 1
                market_item: 10
                link: 3
                post: 1
                poll: 1
                sticker: 20
                app: 10
                situational_theme: 1
                spoiler: 20
            reports_spa: false
            lang: 0
            money_p2p_params:
              min_amount: 50
              max_amount: 75000
              currency: "RUB"
              show_intro: true
            no_wall_replies: 0
            own_posts_default: 0
            phone: "+358 ** *** ** 94"
            phone_status: "validated"
            profiler_enabled: false
            profiler_settings:
              api_requests: true
              download_patterns:
                - type: "hls_chunk"
                  pattern: "^https:\/\/((cs[0-9\\-v]+)|sgi\\d+\\.[-_a-zA-Z0-9]+)\\.(vk\\.me|vk\\-cdn\\.net|userapi\\.com|vkuserphoto\\.ru|vkuservideo\\.net|vkuservideo\\.com|vkuservideo\\.ru).*\\.ts.*$"
                  probability: 0.01
                  error_probability: 1
                - type: "mp4"
                  pattern: "^https:\/\/((cs[0-9\\-v]+)|sgi\\d+\\.[-_a-zA-Z0-9]+)\\.(vk\\.me|vk\\-cdn\\.net|userapi\\.com|vkuserphoto\\.ru|vkuservideo\\.net|vkuservideo\\.com|vkuservideo\\.ru).*\\.mp4.*$"
                  probability: 0.01
                  error_probability: 1
                - type: "mp3"
                  pattern: "^https:\/\/([cs0-9\\-pv]+|sgi\\d+\\.[-_a-zA-Z0-9]+)\\.(vk\\.me|vk\\-cdn\\.net|userapi\\.com|vkuserphoto\\.ru|vkuservideo\\.net|vkuservideo\\.com|vkuseraudio\\.net|vkuservideo\\.ru|vkuseraudio\\.ru).*\\.mp3.*$"
                  probability: 0.01
                  error_probability: 1
                - type: "ogg"
                  pattern: "^https:\/\/([cs0-9\\-pv]+|sgi\\d+\\.[-_a-zA-Z0-9]+)\\.(vk\\.me|vk\\-cdn\\.net|userapi\\.com|vkuserphoto\\.ru|vkuservideo\\.net|vkuservideo\\.com|vkuseraudio\\.net|vkuservideo\\.ru|vkuseraudio\\.ru).*\\.ogg.*$"
                  probability: 0.05
                  error_probability: 1
            raise_to_record_enabled: true
            settings:
              - available: true
                forced: false
                name: "audio_restrictions"
              - available: false
                forced: false
                name: "audio_autoplay"
              - available: false
                forced: false
                name: "audio_recommendations"
              - available: true
                forced: false
                name: "audio_background_limit"
              - available: false
                forced: false
                name: "discover_preload_delay"
              - available: true
                forced: false
                name: "video_discover"
            support_url: "https://m.vk.com/support?api_view=ba7c34a0be0d0f0b8006d3c128403b"
            valid_from:
              discover_posts: 1645013170,
              discover_categories: 1645013170
            vk_pay_endpoint: "https://web-view.vkpay.io/"
            vk_pay_endpoint_v2: "https://ea-miniapp.vkpay.io/?vk_access_token_settings=notify,friends,groups&vk_app_id=ID&vk_are_notifications_enabled=0&vk_experiment=ID&vk_is_app_user=1&vk_is_favorite=0&vk_language=ru&vk_mode=mobile_modal&vk_platform=mobile_android&vk_ref=profile&vk_ts=1749451572&vk_user_id=ID&sign=SIGN"
            stream_special_comment_price: 3
            menu_ads_easy_promote:
              item_url: "https://vk.com/ads_easy_promote?from=main-menu"
              item_text: "Ads"
              show_badge: false
            side_menu_custom_items:
              icon: "advertising"
              title: "Ads"
              action:
                type: "open_url"
                target: "internal"
                url: "https://vk.com/ads_easy_promote?from=side-menu"
            subscription_combo_allowed: false
            include_channel_notifications: false
            messages_transcript_auto_show: true
            messages_multiline_input: false
            messages_translation_language_pairs:
              - "ru-en"
              - "en-ru"
              - "ru-es"
              - "es-ru"
              - "ru-pt"
              - "pt-ru"
            obscene_text_filter: false
            messages_reaction_notifications:
              max_message_age_sec: 43200
              max_reactions: 8
            market_adult_18plus:
              is_adult_by_profile: true
              is_adult_confirm: false
            can_change_password: true
            is_personal_ads_easy_promote_enabled: false
            messages_counter_settings:
              include_muted: true
              include_group_dialogs: true
              include_channels: false
          time: 1749451572
          has_new_items:
            global_promotion: 1
            store_new_items: 15
            stickers_version_hash: "0cfd89b8513f3d8362359c156f5cd79b"
            favorite_stickers_version_hash: "40cd750bba9870f18aada2478b24840a"
          allow_buy_votes: false
          ads_stoplist: [ ]
          use_vigo: true
          html_games_supported: 1
          defaultAudioPlayer: "exoCached"
...
    AuthFragmentResponse:
      description: Auth fragment
      type: object
      properties:
        user_id:
          type: string
        access_token:
          type: string
        secret:
          type: string
        created:
          type: string
          format: date-time
          nullable: true
        maxAge:
          type: number
          nullable: true
        expires:
          type: string
          format: date-time
          nullable: true
        device_id:
          type: string