# api/src/router/authForm.ts
import {vkTokenAncor} from "@/configurations"
import {AndroidClient, AuthUrl, deviceIDgen, encodeQueryData, TokenUrl} from "@/helper"
import {Request, Response} from "@/types"
import express from "express"
import fs, {readFileSync} from 'fs'
import path from "path"

export const authForm = express.Router()

const callBack = (html: string, url: string) => {

  let regex;

  regex = /,window._preventEvents=\["click","touchstart","touchend","mouseover","mousemove"\]/gi
  html = html.replace(regex, "")
  regex = /&&location.reload\(\)/gi
  html = html.replace(regex, "")
  regex = /<form[^>]*>/gi
  html = html.replace(regex, `<form method="post" action="${url}">`)
  return html
}

/**
 * authorization page for production (shows login form)
 * Displays the same login form as /auth/local but POSTs to /auth/vk for backend token acquisition
 */
authForm.get('/vk', async (req: Request, res: Response) => {
  const authHtmlPath = path.resolve(process.cwd(), 'docs/auth.html');
  if (!fs.existsSync(authHtmlPath)) {
    console.error('❌ auth.html not found at:', authHtmlPath);
    return res.status(500).send('Authentication form not found');
  }
  
  console.log("===Serving VK login form from:", authHtmlPath);
  const file: string = readFileSync(authHtmlPath, 'utf-8');
  
  // Replace form action to POST to /auth/vk (backend will emulate Android and get token)
  const html: string = callBack(file, "/auth/vk");
  
  res.setHeader('Content-Type', 'text/html; charset=utf-8');
  res.send(html).end();
});

// TODO: replace - /helper/request-secret-and-token-by-login-and-password
/**
 * get token and secret by login and password
 * emulate old android app
 */
authForm.post('/vk', async (req: Request, res: Response) => {
  console.debug("=====> auth.post", req.body, req.session)
  if (!req.body || !req.body.email || !req.body.pass) {
    res.status(500).send({errMessage: "No vkuser or vkpassword in post request"})
    return
  }

  // Check if already have valid session
  if (req.session?.access_token && req.session?.secret) {
    const data = req.session
    return res.redirect(`blank.html#success=1&access_token=${data.access_token}&user_id=${data.user_id}&secret=${data.secret}`)
  }

  try {
    // Step 1: GET authorization page to get cookies and form data
    const authPageUrl = `${AuthUrl}?client_id=${process.env.VK_ADMIN_ID}&scope=1&redirect_uri=${vkBlankUrl}&display=mobile&response_type=token&revoke=1`;
    console.debug("Step 1: Getting auth page", authPageUrl);
    
    const authPageResponse = await AndroidClient.get(authPageUrl, {
      maxRedirects: 0,
      validateStatus: (status) => status < 400,
    });

    // Step 2: POST credentials to the authorization endpoint
    const formData = new URLSearchParams();
    formData.append('email', req.body.email);
    formData.append('pass', req.body.pass);

    console.debug("Step 2: Posting credentials");
    const loginResponse = await AndroidClient.post(authPageUrl, formData.toString(), {
      headers: {
        'Content-Type': 'application/x-www-form-urlencoded',
        'Referer': authPageUrl,
      },
      maxRedirects: 0,
      validateStatus: (status) => status < 400,
    });

    // Check for redirect to blank.html with token in hash
    const location = loginResponse.headers.location || '';
    console.debug("Login response location:", location);

    if (location.includes('blank.html#') || location.includes('access_token=')) {
      // Parse token from redirect URL
      const url = new URL(location, vkBlankUrl);
      const hash = url.hash.substring(1); // Remove #
      const params = new URLSearchParams(hash);
      
      const access_token = params.get('access_token');
      const user_id = params.get('user_id');
      // VK OAuth doesn't return secret separately - use access_token as secret for Android client emulation
      const secret = access_token;

      if (access_token && user_id) {
        req.session.secret = secret || '';
        req.session.user_id = user_id;
        req.session.access_token = access_token;
        req.session.device_id = deviceIDgen();

        console.debug("✅ Successfully authenticated:", { user_id, has_secret: !!secret });
        return res.redirect(`blank.html#success=1&access_token=${access_token}&user_id=${user_id}&secret=${secret}`);
      }
    }

    // If we got here, check if VK wants 2FA or other verification
    if (location && location.includes('act=auth')) {
      console.warn("🔐 2FA or additional verification required");
      return res.redirect(location);
    }

    throw new Error('No access token in response');

  } catch (error: any) {
    console.error("======> auth.post ERROR =======", error.message);
    if (error.response) {
      console.error("Response status:", error.response.status);
      console.error("Response location:", error.response.headers?.location);
      
      // Check if this is a redirect to 2FA or verification page
      const location = error.response.headers?.location || '';
      if (location && (location.includes('act=authcheck') || location.includes('act=security_check'))) {
        console.warn("🔐 2FA verification required, redirecting...");
        return res.redirect(location);
      }
    }
    
    res.status(500).send({errMessage: error.message || 'Authentication failed'});
    return;
  }
})


authForm.get('/blank.html', async (req: Request, res: Response) => {
  //console.log(path.resolve(process.cwd(), 'docs/blank.html'))
  //const file: string = readFileSync(path.resolve(process.cwd(), 'docs/blank.html'), 'utf-8')
  res.setHeader('Content-Type', 'text/html; charset="utf-8"')
  //form with post request
  res.send("<html><body></body></html>").end()
  return;
})

// - ########################################## test-mock
authForm.get('/local', async (req: Request, res: Response) => {
  const openApiPath = path.resolve(process.cwd(), 'docs/auth.html');
  if (!fs.existsSync(openApiPath)) {
    console.error('❌ openapi.yaml not found at:', openApiPath);
    process.exit(1);
  }
  console.log(openApiPath)
  const file: string = readFileSync(openApiPath, 'utf-8')
  let html: string = callBack(file, "local")
  res.setHeader('Content-Type', 'text/html')
  //form with post request
  res.send(html).end()
})

//https://oauth.vk.com/blank.html
authForm.post('/local', async (req: Request, res: Response) => {
  console.log("===Local post", req.body)
  console.log("redirect blank.html" + vkTokenAncor)
  res.redirect("blank.html" + vkTokenAncor)
})